requirements/requirements-test.txt
-r requirements.txt

pytest==8.3.5
pytest-cov==6.1.1
pytest-django==4.11.1
